Fix pinout for 10118194-0001LF add 10103592-0001LF
Pinout for the 10118194-0001LF was wrong (figured this out with a PCB in
my hand). Either it has been confused with the 10103592-0001LF (same
connector but flipped on it's head) or someone made the mistake of
looking at a male pinout for a female part. Either way, this should fix
it.